+++ test/Unit/Ads/ProgrammaticTrackingServiceTest.ts
@@ -3,7 +3,8 @@ import {
     ProgrammaticTrackingService,
     IProgrammaticTrackingData,
     AdmobMetric,
-    TimingMetric
+    TimingMetric,
+    LoadMetric
 } from 'Ads/Utilities/ProgrammaticTrackingService';
 import { assert } from 'chai';
 import { Platform } from 'Core/Constants/Platform';
@@ -18,18 +19,24 @@ describe('ProgrammaticTrackingService', () => {
     let programmaticTrackingService: ProgrammaticTrackingService;
     let osVersionStub: sinon.SinonStub;
     let sdkVersionStub: sinon.SinonStub;
+    let gameIdStub: sinon.SinonStub;
     let postStub: sinon.SinonStub;
     let platform: Platform;
+    const osVersion = '11.2.1';
+    const sdkVersion = '2300';
 
     beforeEach(() => {
         const request = sinon.createStubInstance(RequestManager);
         const clientInfo = sinon.createStubInstance(ClientInfo);
         const deviceInfo = sinon.createStubInstance(DeviceInfo);
         platform = Platform.ANDROID;
-        programmaticTrackingService = new ProgrammaticTrackingService(platform, request, clientInfo, deviceInfo);
+        programmaticTrackingService = new ProgrammaticTrackingService(platform, request, clientInfo, deviceInfo, 'us');
         osVersionStub = deviceInfo.getOsVersion;
         sdkVersionStub = clientInfo.getSdkVersionName;
+        gameIdStub = clientInfo.getGameId;
         postStub = request.post;
+        osVersionStub.returns(osVersion);
+        sdkVersionStub.returns(sdkVersion);
         postStub.resolves({
             url: 'test',
             response: 'test',
@@ -39,16 +46,8 @@ describe('ProgrammaticTrackingService', () => {
     });
 
     describe('reportErrorEvent', () => {
-
-        const osVersion = '11.2.1';
-        const sdkVersion = '2.3.0';
         const adType = 'test';
         const seatId = 1234;
-
-        beforeEach(() => {
-            osVersionStub.returns(osVersion);
-            sdkVersionStub.returns(sdkVersion);
-        });
 
         const tagBuilder = [
             `ads_sdk2_plt:${Platform[Platform.ANDROID]}`,
@@ -92,7 +91,7 @@ describe('ProgrammaticTrackingService', () => {
             }
         }];
         tests.forEach((t) => {
-            it(`should send "${t.expected.metrics![0].name}" when "${t.input}" is passed in`, () => {
+            it(`should send "${t.expected.metrics[0].name}" when "${t.input}" is passed in`, () => {
                 const promise = programmaticTrackingService.reportErrorEvent(t.input, adType, seatId);
                 sinon.assert.calledOnce(postStub);
                 assert.equal(postStub.firstCall.args.length, 3);
@@ -105,6 +104,11 @@ describe('ProgrammaticTrackingService', () => {
     });
 
     describe('reportMetricEvent', () => {
+
+        beforeEach(() => {
+            gameIdStub.returns('1111');
+        });
+
         const tests: {
             input: AdmobMetric;
             expected: IProgrammaticTrackingData;
@@ -136,7 +140,60 @@ describe('ProgrammaticTrackingService', () => {
             }
         }];
         tests.forEach((t) => {
-            it(`should send "${t.expected.metrics![0].name}" when "${t.input}" is passed in`, () => {
+            it(`should send "${t.expected.metrics[0].name}" when "${t.input}" is passed in`, () => {
+                const promise = programmaticTrackingService.reportMetricEvent(t.input);
+                sinon.assert.calledOnce(postStub);
+                assert.equal(postStub.firstCall.args.length, 3);
+                assert.equal(postStub.firstCall.args[0], 'https://sdk-diagnostics.prd.mz.internal.unity3d.com/v1/metrics');
+                assert.equal(postStub.firstCall.args[1], JSON.stringify(t.expected));
+                assert.deepEqual(postStub.firstCall.args[2], [['Content-Type', 'application/json']]);
+                return promise;
+            });
+        });
+    });
+
+    describe('reportMetricEvent with Zynga WWF Games', () => {
+
+        beforeEach(() => {
+            gameIdStub.returns('2895988');
+        });
+
+        const tests: {
+            input: LoadMetric;
+            expected: IProgrammaticTrackingData;
+        }[] = [{
+            input: LoadMetric.LoadEnabledFill,
+            expected: {
+                metrics: [
+                    {
+                        name: 'load_enabled_fill',
+                        value: 1,
+                        tags: [
+                            'ads_sdk2_mevt:load_enabled_fill',
+                            'ads_sdk2_gid:2895988',
+                            'ads_sdk2_plt:ANDROID'
+                        ]
+                    }
+                ]
+            }
+        }, {
+            input: LoadMetric.LoadEnabledShow,
+            expected: {
+                metrics: [
+                    {
+                        name: 'load_enabled_show',
+                        value: 1,
+                        tags: [
+                            'ads_sdk2_mevt:load_enabled_show',
+                            'ads_sdk2_gid:2895988',
+                            'ads_sdk2_plt:ANDROID'
+                        ]
+                    }
+                ]
+            }
+        }];
+        tests.forEach((t) => {
+            it(`should send "${t.expected.metrics[0].name}" when "${t.input}" is passed in`, () => {
                 const promise = programmaticTrackingService.reportMetricEvent(t.input);
                 sinon.assert.calledOnce(postStub);
                 assert.equal(postStub.firstCall.args.length, 3);
@@ -150,22 +207,18 @@ describe('ProgrammaticTrackingService', () => {
 
     describe('reportTimingEvent', () => {
 
-        const sdkVersion = '2300';
-
         beforeEach(() => {
-            sdkVersionStub.returns(sdkVersion);
+            gameIdStub.returns('1111');
         });
 
         const tests: {
             metric: TimingMetric;
             value: number;
-            country: string;
             path: string;
             expected: IProgrammaticTrackingData;
         }[] = [{
-            metric: TimingMetric.WebviewInitializationTime,
+            metric: TimingMetric.TotalWebviewInitializationTime,
             value: 18331,
-            country: 'us',
             path: '/timing',
             expected: {
                 metrics: [
@@ -181,9 +234,8 @@ describe('ProgrammaticTrackingService', () => {
                 ]
             }
         }, {
-            metric: TimingMetric.WebviewInitializationTime,
+            metric: TimingMetric.TotalWebviewInitializationTime,
             value: -1,
-            country: 'hk',
             path: '/metrics',
             expected: {
                 metrics: [
@@ -198,8 +250,8 @@ describe('ProgrammaticTrackingService', () => {
             }
         }];
         tests.forEach((t) => {
-            it(`should send "${t.expected.metrics![0].name}" with "${t.metric}", "${t.value}", and "${t.country}" is passed in`, () => {
-                const promise = programmaticTrackingService.reportTimingEvent(t.metric, t.value, t.country);
+            it(`should send "${t.expected.metrics[0].name}" with "${t.metric}" and "${t.value}" is passed in`, () => {
+                const promise = programmaticTrackingService.reportTimingEvent(t.metric, t.value);
                 sinon.assert.calledOnce(postStub);
                 assert.equal(postStub.firstCall.args.length, 3);
                 assert.equal(postStub.firstCall.args[0], 'https://sdk-diagnostics.prd.mz.internal.unity3d.com/v1' + t.path);
@@ -207,6 +259,65 @@ describe('ProgrammaticTrackingService', () => {
                 assert.deepEqual(postStub.firstCall.args[2], [['Content-Type', 'application/json']]);
                 return promise;
             });
+        });
+    });
+
+    describe('Batching Events', () => {
+
+        it('should not fire events when no events are batched', () => {
+            return programmaticTrackingService.sendBatchedEvents().then(() => {
+                sinon.assert.notCalled(postStub);
+            });
+        });
+
+        it('should not fire events when negative valued events are batched', () => {
+            programmaticTrackingService.batchEvent(TimingMetric.AdsInitializeTime, -200);
+            return programmaticTrackingService.sendBatchedEvents().then(() => {
+                sinon.assert.notCalled(postStub);
+            });
+        });
+
+        it('should fire events when events are batched', () => {
+            const expected = {
+                metrics: [
+                    {
+                        name: 'uads_core_initialize_time',
+                        value: 999,
+                        tags: [
+                            'ads_sdk2_sdv:2300',
+                            'ads_sdk2_iso:us',
+                            `ads_sdk2_plt:ANDROID`
+                        ]
+                    }, {
+                        name: 'webview_load_to_configuration_complete_time',
+                        value: 100,
+                        tags: [
+                            'ads_sdk2_sdv:2300',
+                            'ads_sdk2_iso:us',
+                            `ads_sdk2_plt:ANDROID`
+                        ]
+                    }
+                ]
+            };
+            programmaticTrackingService.batchEvent(TimingMetric.CoreInitializeTime, 999);
+            programmaticTrackingService.batchEvent(TimingMetric.WebviewLoadToConfigurationCompleteTime, 100);
+            return programmaticTrackingService.sendBatchedEvents().then(() => {
+                sinon.assert.calledOnce(postStub);
+                assert.equal(postStub.firstCall.args.length, 3);
+                assert.equal(postStub.firstCall.args[0], 'https://sdk-diagnostics.prd.mz.internal.unity3d.com/v1/timing');
+                assert.deepEqual(postStub.firstCall.args[1], JSON.stringify(expected));
+                assert.deepEqual(postStub.firstCall.args[2], [['Content-Type', 'application/json']]);
+                // tslint:disable-next-line:no-string-literal
+                assert.deepEqual(programmaticTrackingService['_batchedEvents'], []);
+            });
+        });
+
+        it('should fire events when 10 events are reached', () => {
+            for (let i = 0; i < 10; i++) {
+                sinon.assert.notCalled(postStub);
+                programmaticTrackingService.batchEvent(TimingMetric.TotalWebviewInitializationTime, 200);
+            }
+            sinon.assert.calledOnce(postStub);
         });
     });
 });